fix: avoid comparing destroyed bytes
Make `operator ==` symmetric by also checking that the argument bytes
are not destroyed before reading them.

In the `operator ==` implementation for `SecretKeyData` the receiver is
first checked for destroyed bytes which are considered unequal to any
other key. The argument was not checked for destroyed bytes, so in an
expression `notDestroyed == destroyed` the `bytes` getter on the
argument would throw a `StateError`.

This exception is masked in tests by the `equals` matcher which
currently treats an exception from `operator ==` the same as if it had
returned `false`. In an upcoming release this exception will stop being
masked and will start to cause test failures.